+++ sample-compose/src/main/java/org/koin/sample/androidx/compose/di/appModule.kt
@@ -1,11 +1,12 @@
 package org.koin.sample.androidx.compose.di
 
-import org.koin.androidx.viewmodel.dsl.viewModel
+import org.koin.androidx.viewmodel.dsl.viewModelOf
+import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
 import org.koin.sample.androidx.compose.data.UserRepository
 import org.koin.sample.androidx.compose.viewmodel.UserViewModel
 
 val appModule = module {
-    viewModel { UserViewModel(get()) }
-    single { UserRepository() }
+    viewModelOf(::UserViewModel)
+    singleOf(::UserRepository)
 }